Added in Schur_Complement first try.
This includes a test (in the "tests" folder) which is written in python.
The "SweeperSchurBoundary.cc" file contains the current version of my attempt at
the Schur algorithm. It is effectively the same as PBJ, except that it ignores a majority
of the psi values on the interiors of the domains and instead only iterates based on the boundary
values. It runs and returns the same values as Sweeper2 and SweeperPBJ (based on the test I wrote),
but it takes more iterations than PBJ for some reason. Currently, the output it is saving for the tests
is text files (to make it human-readable while I was troubleshooting). I need to make them binary files to
save space and add a toggle (probably in input.deck) to turn writing results on/off.