Serial Version is running, but gives wrong answer.
Made some headway on why the answer is incorrect also. I think the vector "b" (the RHS) is currently set to 0, and so the Krylov solver quickly converges to zero. Tried setting b to the correct values, but ended up with segfaults. Those changes may appear as comments in this commit. The parallel version has an issue with the MPI: all the local and global vectors have some size mismatch thats exactly a factor of g_numThreads.