The cmake configuration now has a -DBUILD_UTILS:BOOL to optionally install the mesh util directory.
The OutputFile name keyword in the input is now a "prefix"
string for the outfiles. The binary angular flux and scalar
flux are written to "prefix".psi and "prefix".phi, while
the cell-average scalar flux is written to the file
"prefix".phi.txt

Added a simple test script call run.sh.